Controller API packages restructure.
