+++ mipengine/controller/resources/pathologies_metadata/dementia.json
