+++ mipengine/controller/resources/pathologies_metadata/mentalhealth.json
