+++ mipengine/controller/resources/worker_catalogue.json
